+++ kong/plugins/opentelemetry/schema.lua
@@ -1,5 +1,7 @@
 local typedefs = require "kong.db.schema.typedefs"
 local Schema = require "kong.db.schema"
+local deprecation = require("kong.deprecation")
+local QUEUE_CONFIGURATION_SCHEMA = require("kong.tools.queue").configuration_schema
 
 local function custom_validator(attributes)
   for _, v in pairs(attributes) do
@@ -44,11 +46,31 @@ return {
           },
         } },
         { resource_attributes = resource_attributes },
-        { batch_span_count = { type = "integer", required = true, default = 200 } },
-        { batch_flush_delay = { type = "integer", required = true, default = 3 } },
+        { batch_span_count = { type = "integer", required = true, default = 200 } }, -- deprecated, use queue.batch_max_size
+        { batch_flush_delay = { type = "integer", required = true, default = 3 } }, -- deprecated, use queue.max_delay
+          { queue = {
+            type = "record",
+            fields = QUEUE_CONFIGURATION_SCHEMA,
+          }},
         { connect_timeout = typedefs.timeout { default = 1000 } },
         { send_timeout = typedefs.timeout { default = 5000 } },
         { read_timeout = typedefs.timeout { default = 5000 } },
+      },
+      entity_checks = {
+        { custom_entity_check = {
+          field_sources = { "batch_span_count", "batch_flush_delay" },
+          fn = function(entity)
+            if entity.batch_span_count then
+              deprecation("batch_span_count is deprecated, please use queue.batch_max_size instead",
+                          { after = "4.0", })
+            end
+            if entity.batch_flush_delay then
+              deprecation("batch_flush_delay is deprecated, please use queue.batch_max_size instead",
+                          { after = "4.0", })
+            end
+            return true
+          end
+        } },
       },
     }, },
   },