+++ spec/01-unit/01-db/01-schema/11-declarative_config/02-process_auto_fields_spec.lua
@@ -175,6 +175,15 @@ describe("declarative config: process_auto_fields", function()
                 queue_size = 1,
                 retry_count = 10,
                 timeout = 10000,
+                queue = {
+                  batch_max_size = 1,
+                  capacity = 10000,
+                  max_delay = 1,
+                  max_idle_time = 60,
+                  max_retry_delay = 60,
+                  max_retry_time = 60,
+                  poll_time = 1,
+                },
               }
             },
           }
@@ -227,6 +236,15 @@ describe("declarative config: process_auto_fields", function()
                 queue_size = 1,
                 retry_count = 10,
                 timeout = 10000,
+                queue = {
+                  batch_max_size = 1,
+                  capacity = 10000,
+                  max_delay = 1,
+                  max_idle_time = 60,
+                  max_retry_delay = 60,
+                  max_retry_time = 60,
+                  poll_time = 1,
+                },
               }
             },
           }
@@ -339,6 +357,15 @@ describe("declarative config: process_auto_fields", function()
                       queue_size = 1,
                       retry_count = 10,
                       timeout = 10000,
+                      queue = {
+                        batch_max_size = 1,
+                        capacity = 10000,
+                        max_delay = 1,
+                        max_idle_time = 60,
+                        max_retry_delay = 60,
+                        max_retry_time = 60,
+                        poll_time = 1,
+                      },
                     }
                   },
                 }
@@ -654,6 +681,15 @@ describe("declarative config: process_auto_fields", function()
                           queue_size = 1,
                           retry_count = 10,
                           timeout = 10000,
+                          queue = {
+                            batch_max_size = 1,
+                            capacity = 10000,
+                            max_delay = 1,
+                            max_idle_time = 60,
+                            max_retry_delay = 60,
+                            max_retry_time = 60,
+                            poll_time = 1,
+                          },
                         }
                       }
                     }